Fix namespace issues.

+++ trunk/fftw++/cconv3.cc
@@ -30,7 +30,7 @@ bool Direct=false, Implicit=true, Explicit=false, Pruned=false;
 
 inline void init(array3<Complex>& f, array3<Complex>& g, unsigned int M=1) 
 {
-  double factor=1.0/sqrt(M);
+  double factor=1.0/sqrt((double) M);
   for(unsigned int s=0; s < M; ++s) {
     double S=sqrt(1.0+s);
     double ffactor=S*factor;
