+++ trunk/fftw++/conv.cc
@@ -28,7 +28,7 @@ bool Direct=false, Implicit=true, Explicit=false, Test=false;
 inline void init(Complex *f, Complex *g, unsigned int M=1) 
 {
   unsigned int Mm=M*m;
-  double factor=1.0/sqrt(M);
+  double factor=1.0/sqrt((double) M);
   for(unsigned int i=0; i < Mm; i += m) {
     double ffactor=(1.0+i)*factor;
     double gfactor=1.0/(1.0+i)*factor;
